Rework RLHF class
