+++ examples/nanoChatGPT/config/train_rlhf.yaml
@@ -13,7 +13,7 @@ eval_iters: 200
 eval_only: False # if True, script exits right after the first eval
 always_save_checkpoint: True # if True, always save a checkpoint after each eval
 init_from: resume # 'scratch' or 'resume' or 'gpt2*'
-init_multihead_from: resume
+init_reward_from: resume
 # wandb:
 wandb_log: False # disabled by default
 wandb_project: rlhf # 'gpt2'
